requirements.txt
fastapi==0.112.2
uvicorn==0.30.6
pymongo==4.8.0
python-dotenv==1.0.1
openai==1.44.0
langchain==0.2.14
langchain-openai==0.1.22
pydantic==2.9.2
